requirements.txt
Flask
numpy
scikit-image
pillow
roboflow
svgutils